# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cuarto Arte API - Gestión de eventos musicales y artísticos
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /musicians:
    get:
      operationId: getMusicians
      tags: [musicians]
      summary: List all musicians
      responses:
        "200":
          description: List of musicians
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Musician"
    post:
      operationId: createMusician
      tags: [musicians]
      summary: Create a new musician
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMusician"
      responses:
        "201":
          description: Musician created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Musician"
  /musicians/{id}:
    get:
      operationId: getMusician
      tags: [musicians]
      summary: Get a musician by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Musician found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Musician"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateMusician
      tags: [musicians]
      summary: Update a musician
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMusician"
      responses:
        "200":
          description: Musician updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Musician"
    delete:
      operationId: deleteMusician
      tags: [musicians]
      summary: Delete a musician
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Musician deleted
...
  /payments:
    get:
      operationId: getPayments
      tags: [payments]
      summary: List all payments
      parameters:
        - name: eventId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a new payment or advance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePayment"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    put:
      operationId: updatePayment
      tags: [payments]
      summary: Update a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePayment"
      responses:
        "200":
          description: Payment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    delete:
      operationId: deletePayment
      tags: [payments]
      summary: Delete a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Payment deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Musician:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        instruments:
          type: string
        specialty:
          type: string
          nullable: true
        rate:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - phone
        - instruments
        - createdAt
    CreateMusician:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        instruments:
          type: string
        specialty:
          type: string
          nullable: true
        rate:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - name
        - phone
        - instruments
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        eventTitle:
          type: string
          nullable: true
        amount:
          type: number
        type:
          type: string
          enum: [anticipo, pago_total, pago_parcial]
        method:
          type: string
          enum: [efectivo, transferencia, cheque, otro]
        date:
          type: string
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - eventId
        - amount
        - type
        - method
        - date
        - createdAt
    CreatePayment:
      type: object
      properties:
        eventId:
          type: integer
        amount:
          type: number
        type:
          type: string
          enum: [anticipo, pago_total, pago_parcial]
        method:
          type: string
          enum: [efectivo, transferencia, cheque, otro]
        date:
          type: string
        notes:
          type: string
          nullable: true
      required:
        - eventId
        - amount
        - type
        - method
        - date
...