# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cuarto Arte API - Gestión de eventos musicales y artísticos
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentMethod } from "./paymentMethod";
import type { PaymentType } from "./paymentType";

export interface Payment {
  id: number;
  eventId: number;
  eventTitle?: string | null;
  amount: number;
  type: PaymentType;
  method: PaymentMethod;
  date: string;
  notes?: string | null;
  createdAt: string;
}